requirements.txt
streamlit>=1.25
streamlit-authenticator>=0.2.2
python-dotenv>=1.0
pandas>=2.0
numpy>=1.24
pymongo>=4.5
bcrypt>=4.0
transformers>=4.38
torch>=2.2
sentencepiece>=0.2
fpdf2>=2.7.4
PyPDF2>=3.0
python-docx>=0.8
scikit-learn>=1.3
seaborn
